requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
torch==2.7.0
torchvision==0.22.0
fastai>=2.7.13
matplotlib==3.8.0
pandas==2.1.1
numpy==1.26.0
